requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydot==3.0.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
smmap==5.0.1
tzdata==2024.2
urllib3==2.4.0
